requirements.txt
sqlalchemy
quart
asyncmy
hypercorn
python-dotenv
pyjwt
pyarrow
mysqlclient
pymysql